refactor: Move GlobalStaffFactory helper to common/
