+++ lms/djangoapps/courseware/tests/test_view_authentication.py
@@ -9,10 +9,10 @@ from unittest.mock import patch
 import pytz
 from django.urls import reverse
 
+from common.djangoapps.student.tests.factories import GlobalStaffFactory
 from lms.djangoapps.courseware.access import has_access
 from lms.djangoapps.courseware.tests.factories import (
     BetaTesterFactory,
-    GlobalStaffFactory,
     InstructorFactory,
     OrgInstructorFactory,
     OrgStaffFactory,
