+++ openedx/tests/xblock_integration/test_crowdsource_hinter.py
@@ -9,7 +9,7 @@ import simplejson as json
 from django.conf import settings
 from django.urls import reverse
 
-from lms.djangoapps.courseware.tests.factories import GlobalStaffFactory
+from common.djangoapps.student.tests.factories import GlobalStaffFactory
 from lms.djangoapps.courseware.tests.helpers import LoginEnrollmentTestCase
 from openedx.core.lib.url_utils import quote_slashes
 from xmodule.modulestore.tests.django_utils import SharedModuleStoreTestCase
